# Improve SAR performance