@zegerius 2019-09-04
Thank you @gramhagen for that very elegant improvement. I have incorporate your suggested changes. All tests pass. 

Also ran a quick performance comparison and I don't seen any difference compared to our suggested solution (at least not more than one std dev):
`remove_seen = True`: 18.2 ms ± 1.95 ms (from 17.5 ms ± 1.06 ms)
`remove_seen = False`: 20.4 ms ± 2.35 ms (from 17.9 ms ± 1.7 ms)